[refactor] 전체적인 리팩토링 진행 - AnswerEntity 상속관계로 분리하여 추후 확장성 염두하여 설계 - Answer, Survey 모듈 분리하여 의존성 관계 일부 해결 - Validator 모듈 개발(전략적 설계 진행) - request dto common -> core로 변경 - 옵션 타입에 따라 다른 request, response dto 개발 및 적용
