+++ survey-core/src/main/java/org/brinst/surveycore/answer/validator/AnswerValidator.java
@@ -1,0 +1,9 @@
+package org.brinst.surveycore.answer.validator;
+
+import java.util.List;
+
+import org.brinst.surveycore.survey.entity.SurveyQuestion;
+
+public interface AnswerValidator {
+	void validate(SurveyQuestion question, List<String> answers);
+}
