--- survey-core/src/main/java/org/brinst/surveycore/entity/SurveyOption.java
@@ -1,26 +1,0 @@
-package org.brinst.surveycore.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class SurveyOption {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-	private String option;
-	@ManyToOne
-	private SurveyQuestion surveyQuestion;
-
-	public SurveyOption(String option, SurveyQuestion surveyQuestion) {
-		this.option = option;
-		this.surveyQuestion = surveyQuestion;
-	}
-}
