--- survey-core/src/main/java/org/brinst/surveycore/mapper/SurveyDTOMapper.java
@@ -1,10 +1,0 @@
-package org.brinst.surveycore.mapper;
-
-import org.brinst.surveycommon.dto.SurveyDTO;
-import org.brinst.surveycore.entity.Survey;
-
-public class SurveyDTOMapper {
-	// public static SurveyDTO.ResDTO toDTO(Survey survey) {
-	//
-	// }
-}
