--- survey-core/src/main/java/org/brinst/surveycore/repository/SurveyRepository.java
@@ -1,10 +1,0 @@
-package org.brinst.surveycore.repository;
-
-import org.brinst.surveycore.entity.Survey;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface SurveyRepository extends JpaRepository<Survey, Long> {
-
-}
