--- survey-core/src/main/java/org/brinst/surveycore/repository/SurveyVersionRepository.java
@@ -1,9 +1,0 @@
-package org.brinst.surveycore.repository;
-
-import org.brinst.surveycore.entity.Survey;
-import org.brinst.surveycore.entity.SurveyVersion;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SurveyVersionRepository extends JpaRepository<SurveyVersion, Long> {
-	SurveyVersion findByVersionAndSurvey(int version, Survey survey);
-}
