+++ survey-core/src/main/java/org/brinst/surveycore/survey/repository/SurveyRepository.java
@@ -1,0 +1,10 @@
+package org.brinst.surveycore.survey.repository;
+
+import org.brinst.surveycore.survey.entity.Survey;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SurveyRepository extends JpaRepository<Survey, Long> {
+
+}
